requirements.txt
anthropic>=0.40.0
langchain>=0.3.0
langchain-text-splitters>=0.3.0
langchain-community>=0.3.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
rank-bm25>=0.2.2
PyMuPDF>=1.24.0
langdetect>=1.0.9
streamlit>=1.40.0
python-dotenv>=1.0.0
numpy>=1.26.0
ollama>=0.3.0
tqdm>=4.66.0
